remove internal Facebook CDN filename from output filenames
This may be re-implemented as an option in a future version of the
addon, but for now, it appears that for almost all use cases, nobody
needs to know the long string of digits that Facebook names its photos
with.

This also significantly reduces the chance of a file being saved with a
path that's too long.