improved handling of album folder naming and existing album folders
Many users have been annoyed by albums whose folders already exist being skipped
especially when the desired action is to either update the album, or when the
profile has multiple albums with the same name.

Initial solution idea was to offer three options for handling this condition:

1. merge and update the album
2. retry creating album folder with an incremental name (" (2)", " (3)", ...)
3. skip downloading the album (the existing behaviour)

However, this approach would fall apart when a user tries to update multiple
albums with the same given name. The user would initially choose option 2,
resulting in albums named "Album", "Album (2)", etc. then when updating, choose
option 1. When the user does this, all photos from both albums would end up in
the folder called "Album".

The final chosen solution is to append the unique numeric album ID to the folder
name. This way, the folder names are guaranteed unique, and the second option is
not needed. The only downside to this is that because this tends to add about 20
characters to the folder name, the chance of hitting a 'path too long' error is
increased.

Along with the modified naming behaviour, users can now choose between skipping
or updating an album that appears to already exist.